Create README.md
Experienced and highly efficient secretary with a robust background working 
for numerous institutions over many years. Now transitioned into the tech 
industry as a Data Analyst, bringing a wealth of skills in data management, 
analysis and visualization, known for exceptional typing speed and accuracy, 
keen observational abilities and a meticulous approach to detail. A collaborative 
team player with a passion for research, problem-solving and brainstorming 
innovative solutions. Dedicated to leveraging strong analytical skills and a 
detail-oriented mindset to support data-driven decision-making and contribute 
to organizational success